Allow opt-out of the named-pipe listener using an environment variable
