+++ src/System.Management.Automation/utils/Telemetry.cs
@@ -8,7 +8,6 @@ using System.IO;
 using System.Linq;
 using System.Management.Automation;
 using System.Runtime.InteropServices;
-using System.Security.AccessControl;
 using System.Threading;
 
 using Microsoft.ApplicationInsights;
@@ -177,7 +176,7 @@ namespace Microsoft.PowerShell.Telemetry
         static ApplicationInsightsTelemetry()
         {
             // If we can't send telemetry, there's no reason to do any of this
-            CanSendTelemetry = !GetEnvironmentVariableAsBool(name: _telemetryOptoutEnvVar, defaultValue: false);
+            CanSendTelemetry = !Utils.GetEnvironmentVariableAsBool(name: _telemetryOptoutEnvVar, defaultValue: false);
             if (CanSendTelemetry)
             {
                 s_sessionId = Guid.NewGuid().ToString();
@@ -641,72 +640,6 @@ namespace Microsoft.PowerShell.Telemetry
                         "Az Predictor"
                     };
             }
-        }
-
-        /// <summary>
-        /// Determine whether the environment variable is set and how.
-        /// </summary>
-        /// <param name="name">The name of the environment variable.</param>
-        /// <param name="defaultValue">If the environment variable is not set, use this as the default value.</param>
-        /// <returns>A boolean representing the value of the environment variable.</returns>
-        private static bool GetEnvironmentVariableAsBool(string name, bool defaultValue)
-        {
-            var str = Environment.GetEnvironmentVariable(name);
-            if (string.IsNullOrEmpty(str))
-            {
-                return defaultValue;
-            }
-
-            var boolStr = str.AsSpan();
-
-            if (boolStr.Length == 1)
-            {
-                if (boolStr[0] == '1')
-                {
-                    return true;
-                }
-
-                if (boolStr[0] == '0')
-                {
-                    return false;
-                }
-            }
-
-            if (boolStr.Length == 3 &&
-                (boolStr[0] == 'y' || boolStr[0] == 'Y') &&
-                (boolStr[1] == 'e' || boolStr[1] == 'E') &&
-                (boolStr[2] == 's' || boolStr[2] == 'S'))
-            {
-                return true;
-            }
-
-            if (boolStr.Length == 2 &&
-                (boolStr[0] == 'n' || boolStr[0] == 'N') &&
-                (boolStr[1] == 'o' || boolStr[1] == 'O'))
-            {
-                return false;
-            }
-
-            if (boolStr.Length == 4 &&
-                (boolStr[0] == 't' || boolStr[0] == 'T') &&
-                (boolStr[1] == 'r' || boolStr[1] == 'R') &&
-                (boolStr[2] == 'u' || boolStr[2] == 'U') &&
-                (boolStr[3] == 'e' || boolStr[3] == 'E'))
-            {
-                return true;
-            }
-
-            if (boolStr.Length == 5 &&
-                (boolStr[0] == 'f' || boolStr[0] == 'F') &&
-                (boolStr[1] == 'a' || boolStr[1] == 'A') &&
-                (boolStr[2] == 'l' || boolStr[2] == 'L') &&
-                (boolStr[3] == 's' || boolStr[3] == 'S') &&
-                (boolStr[4] == 'e' || boolStr[4] == 'E'))
-            {
-                return false;
-            }
-
-            return defaultValue;
         }
 
         /// <summary>
